follow-up for bug 1138553, Tree.get_or_create needs l10n, use get and create independently
If a tree changes the l10n repos, we want to catch this in
the tree setup. Thus, we can't specify it right away.
So instead of get_or_create, we get first, and if it's not found,
create a tree with l10n.
If the tree is found, check that l10n is actually the forest we
have from the config.