bug 1341886, fix cross-repo includes for unified repos
This depends on a corresponding fix to compare-locales, which
now takes a parameter to map upstream repos to local clones.
Add that information to the build properties, and have the
InspectorStep pick that up and pass it through to compare-locales.

Also put the compare-locales call into a try: except block to
get logs of python errors there.